update more live-repair-related dependency-checking tests
