+++ upstairs/src/live_repair.rs
@@ -1463,375 +1463,7 @@ mod more_tests {
         Logger::root(slog_term::FullFormat::new(plain).build().fuse(), o!())
     }
 
-    // YYY i think this next batch should maybe move to downstairs also
     // YYY artemis-START
-
-    /// Create and enqueue a full set of repair functions
-    ///
-    /// This gets the required locks, determines the next job IDs, checks
-    /// dependencies, and enqueues all four jobs (close / repair / noop /
-    /// reopen).
-    ///
-    /// The repair job is **actually** enqueued as an `ExtentLiveNoop` as well
-    /// (instead of an `ExtentLiveRepair`, because the latter requires more
-    /// infrastructure to be set up in the `Downstairs`; specifically,
-    /// `create_and_enqueue_repair_io` expects `Downstairs::repair_info` to be
-    /// valid.
-    ///
-    /// This function creates the same dependency chain as `repair_extent`,
-    /// which is actually used in a real system, but it doesn't wait for repairs
-    /// to complete before enqueuing stuff.
-    async fn create_and_enqueue_repair_ops(up: &mut Upstairs, eid: u64) {
-        let mut gw = up.guest.guest_work.lock().await;
-        let ds = &mut up.downstairs;
-        ds.clients[ClientId::new(1)].extent_limit =
-            Some(eid.try_into().unwrap());
-
-        let (extent_repair_ids, deps) = ds.get_repair_ids(eid);
-
-        let gw_close_id = gw.next_gw_id();
-        let _close_brw = ds.create_and_enqueue_close_io(
-            &mut gw,
-            eid,
-            4, // gen
-            deps,
-            extent_repair_ids.close_id,
-            gw_close_id,
-            ClientId::new(0),    // source downstairs
-            &[ClientId::new(1)], // repair downstairs
-        );
-        let gw_repair_id = gw.next_gw_id();
-        let _repair_brw = ds.create_and_enqueue_noop_io(
-            &mut gw,
-            vec![extent_repair_ids.close_id],
-            extent_repair_ids.repair_id,
-            gw_repair_id,
-        );
-        let gw_noop_id = gw.next_gw_id();
-        let _noop_brw = ds.create_and_enqueue_noop_io(
-            &mut gw,
-            vec![extent_repair_ids.repair_id],
-            extent_repair_ids.noop_id,
-            gw_noop_id,
-        );
-        let gw_reopen_id = gw.next_gw_id();
-        let _reopen_brw = ds.create_and_enqueue_reopen_io(
-            &mut gw,
-            eid,
-            vec![extent_repair_ids.noop_id],
-            extent_repair_ids.reopen_id,
-            gw_reopen_id,
-        );
-    }
-
-    // The next section of tests verify dependencies are honored.
-
-    // W is Write
-    // R is read
-    // F is flush
-    // Rp is a Repair
-
-    #[tokio::test]
-    async fn test_live_repair_deps_writes() {
-        // Test that writes on different blocks in the extent are all
-        // captured by the repair at the end.
-        //       block
-        // op# | 0 1 2 | deps
-        // ----|-------|-----
-        //   0 | W     |
-        //   1 |   W   |
-        //   2 |     W |
-        //   3 | RpRpRp| 0,1,2
-        //   4 | RpRpRp| 3
-        //   5 | RpRpRp| 4
-        //   6 | RpRpRp| 5
-
-        let up = start_up_and_repair(ClientId::new(1)).await;
-        // Write operations 0 to 2
-        for i in 0..3 {
-            up.submit_dummy_write(
-                Block::new_512(i),
-                Bytes::from(vec![0xff; 512]),
-                false,
-            )
-            .await;
-        }
-
-        // Repair IO functions assume you have the locks
-        let mut ds = up.downstairs;
-        let eid = 0;
-        ds.clients[ClientId::new(1)].extent_limit =
-            Some(eid.try_into().unwrap());
-        drop(ds);
-
-        create_and_enqueue_repair_ops(&mut up, eid).await;
-
-        let ds = up.downstairs;
-        let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
-
-        assert_eq!(jobs.len(), 7);
-
-        // The three writes don't depend on anything
-        assert!(jobs[0].work.deps().is_empty());
-        assert!(jobs[1].work.deps().is_empty());
-        assert!(jobs[2].work.deps().is_empty());
-        // The repair will have all the previous jobs, as they all are on the
-        // same extent.
-        assert_eq!(
-            jobs[3].work.deps(),
-            &[jobs[0].ds_id, jobs[1].ds_id, jobs[2].ds_id]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_live_repair_deps_reads() {
-        // Test that the following job dependency graph is made:
-        //
-        //       block
-        // op# | 0 1 2 | deps
-        // ----|-------|-----
-        //   0 | R     |
-        //   1 |   R   |
-        //   2 |     R |
-        //   3 | RpRpRp| 0,1,2
-        //   4 | RpRpRp| 3
-        //   5 | RpRpRp| 4
-        //   6 | RpRpRp| 5
-
-        let up = start_up_and_repair(ClientId::new(1)).await;
-
-        // Create read operations 0 to 2
-        for i in 0..3 {
-            up.submit_dummy_read(Block::new_512(i), Buffer::new(512))
-                .await;
-        }
-
-        // Repair IO functions assume you have the locks already
-        let mut ds = up.downstairs;
-        let eid = 0;
-        ds.clients[ClientId::new(1)].extent_limit =
-            Some(eid.try_into().unwrap());
-        drop(ds);
-
-        create_and_enqueue_repair_ops(&mut up, eid).await;
-
-        let ds = up.downstairs;
-        let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
-
-        assert_eq!(jobs.len(), 7);
-
-        // The three reads don't depend on anything
-        assert!(jobs[0].work.deps().is_empty());
-        assert!(jobs[1].work.deps().is_empty());
-        assert!(jobs[2].work.deps().is_empty());
-        // The repair will have all the previous jobs, as they all are on the
-        // same extent.
-        assert_eq!(
-            jobs[3].work.deps(),
-            &[jobs[0].ds_id, jobs[1].ds_id, jobs[2].ds_id]
-        );
-    }
-
-    #[tokio::test]
-    async fn test_live_repair_deps_mix() {
-        // Test that the following job dependency graph is made:
-        //
-        //       block
-        // op# | 0 1 2 | deps
-        // ----|-------|-----
-        //   0 | R     |
-        //   1 |   W   |
-        //   2 | F F F | 0,1
-        //   3 |RpRpRp | 2
-        //   4 |RpRpRp |
-        //   5 |RpRpRp |
-        //   6 |RpRpRp |
-
-        let up = start_up_and_repair(ClientId::new(1)).await;
-        up.submit_dummy_read(Block::new_512(0), Buffer::new(512))
-            .await;
-
-        up.submit_dummy_write(
-            Block::new_512(1),
-            Bytes::from(vec![0xff; 512]),
-            false,
-        )
-        .await;
-
-        up.submit_flush(None, None).await;
-
-        // Repair IO functions assume you have the locks already
-        let mut ds = up.downstairs;
-        let eid = 0;
-        ds.clients[ClientId::new(1)].extent_limit =
-            Some(eid.try_into().unwrap());
-        drop(ds);
-
-        create_and_enqueue_repair_ops(&mut up, eid).await;
-
-        let ds = up.downstairs;
-        let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
-
-        assert_eq!(jobs.len(), 7);
-
-        // The read and the write don't depend on anything
-        assert!(jobs[0].work.deps().is_empty());
-        assert!(jobs[1].work.deps().is_empty());
-        // The flush requires the read and the write
-        assert_eq!(jobs[2].work.deps(), &[jobs[0].ds_id, jobs[1].ds_id]);
-        // The repair will have just the flush
-        assert_eq!(jobs[3].work.deps(), &[jobs[2].ds_id]);
-    }
-
-    #[tokio::test]
-    async fn test_live_repair_deps_repair() {
-        // Basic test for inter-repair dependencies
-        //
-        // This only actually tests our test function
-        // (`create_and_enqueue_repair_ops`), not the actual `repair_extent`,
-        // but is still worthwhile
-        //
-        //       block
-        // op# | 0 1 2 | deps
-        // ----|-------|-----
-        //   0 | RpRpRp|
-        //   1 | RpRpRp| 0
-        //   2 | RpRpRp| 1
-        //   3 | RpRpRp| 2
-        let up = start_up_and_repair(ClientId::new(1)).await;
-        let eid = 0;
-        create_and_enqueue_repair_ops(&mut up, eid).await;
-
-        let ds = up.downstairs;
-        let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
-
-        assert_eq!(jobs.len(), 4);
-
-        assert_eq!(jobs[0].ds_id, JobId(1000));
-        assert!(jobs[0].work.deps().is_empty());
-        assert_eq!(jobs[1].ds_id, JobId(1001));
-        assert_eq!(jobs[1].work.deps(), &[JobId(1000)]);
-        assert_eq!(jobs[2].ds_id, JobId(1002));
-        assert_eq!(jobs[2].work.deps(), &[JobId(1001)]);
-        assert_eq!(jobs[3].ds_id, JobId(1003));
-        assert_eq!(jobs[3].work.deps(), &[JobId(1002)]);
-    }
-
-    #[tokio::test]
-    async fn test_live_repair_deps_repair_write() {
-        // Write after repair depends on the repair
-        //       block
-        // op# | 0 1 2 | deps
-        // ----|-------|-----
-        //   0 | RpRpRp|
-        //   1 | RpRpRp| 0
-        //   2 | RpRpRp| 1
-        //   3 | RpRpRp| 2
-        //   4 |     W | 3
-
-        let up = start_up_and_repair(ClientId::new(1)).await;
-        // Repair IO functions assume you have the locks
-        let mut ds = up.downstairs;
-        let eid = 0;
-        ds.clients[ClientId::new(1)].extent_limit =
-            Some(eid.try_into().unwrap());
-
-        drop(ds);
-
-        create_and_enqueue_repair_ops(&mut up, eid).await;
-
-        up.submit_dummy_write(
-            Block::new_512(2),
-            Bytes::from(vec![0xff; 512]),
-            false,
-        )
-        .await;
-
-        let ds = up.downstairs;
-        let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
-
-        assert_eq!(jobs.len(), 5);
-
-        assert_eq!(jobs[0].ds_id, JobId(1000));
-        assert!(jobs[0].work.deps().is_empty());
-
-        assert_eq!(jobs[4].ds_id, JobId(1004));
-        assert_eq!(jobs[4].work.deps(), &[JobId(1003)]);
-    }
-
-    #[tokio::test]
-    async fn test_live_repair_deps_repair_read() {
-        // Read after repair requires the repair
-        //       block
-        // op# | 0 1 2 | deps
-        // ----|-------|-----
-        //   0 | RpRpRp|
-        //   1 | RpRpRp|
-        //   2 | RpRpRp|
-        //   3 | RpRpRp|
-        //   4 | R     | 3
-        let up = start_up_and_repair(ClientId::new(1)).await;
-        // Repair IO functions assume you have the locks already
-        let mut ds = up.downstairs;
-        let eid = 0;
-        ds.clients[ClientId::new(1)].extent_limit =
-            Some(eid.try_into().unwrap());
-        drop(ds);
-
-        create_and_enqueue_repair_ops(&mut up, eid).await;
-
-        up.submit_dummy_read(Block::new_512(0), Buffer::new(512))
-            .await;
-
-        let ds = up.downstairs;
-
-        let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
-
-        assert_eq!(jobs.len(), 5);
-
-        assert_eq!(jobs[0].ds_id, JobId(1000));
-        assert!(jobs[0].work.deps().is_empty());
-        // The read depends on the last item of the repair
-        assert_eq!(jobs[4].ds_id, JobId(1004));
-        assert_eq!(jobs[4].work.deps(), &[JobId(1003)]);
-    }
-
-    #[tokio::test]
-    async fn test_live_repair_deps_repair_flush() {
-        // Flush after repair requires the flush
-        //       block
-        // op# | 0 1 2 | deps
-        // ----|-------|-----
-        //   0 | RpRpRp|
-        //   1 | RpRpRp|
-        //   2 | RpRpRp|
-        //   3 | RpRpRp|
-        //   4 | F F F | 3
-        let up = start_up_and_repair(ClientId::new(1)).await;
-        // Repair IO functions assume you have the locks already
-        let mut ds = up.downstairs;
-        let eid = 0;
-        ds.clients[ClientId::new(1)].extent_limit =
-            Some(eid.try_into().unwrap());
-        drop(ds);
-
-        create_and_enqueue_repair_ops(&mut up, eid).await;
-
-        up.submit_flush(None, None).await;
-
-        let ds = up.downstairs;
-
-        let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
-
-        assert_eq!(jobs.len(), 5);
-
-        assert_eq!(jobs[0].ds_id, JobId(1000));
-        assert!(jobs[0].work.deps().is_empty());
-        // The flush depends on the repair close operation
-        assert_eq!(jobs[4].ds_id, JobId(1004));
-        assert_eq!(jobs[4].work.deps(), &[JobId(1003)]);
-    }
-
     #[tokio::test]
     async fn test_live_repair_deps_no_overlap() {
         // No overlap, no deps, IO before repair
@@ -1855,7 +1487,7 @@ mod more_tests {
         up.submit_dummy_read(Block::new_512(6), Buffer::new(512))
             .await;
 
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -1881,7 +1513,7 @@ mod more_tests {
         //   4 |       |       |   R   |
         //   5 |  W    |       |       |
         let up = start_up_and_repair(ClientId::new(1)).await;
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         up.submit_dummy_write(
             Block::new_512(2),
@@ -1919,7 +1551,7 @@ mod more_tests {
         let up = start_up_and_repair(ClientId::new(1)).await;
         up.submit_flush(None, None).await;
 
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         up.submit_flush(None, None).await;
 
@@ -1954,11 +1586,11 @@ mod more_tests {
         //   7 |       | RpRpRp|
         //   8 |       | RpRpRp|
         let up = start_up_and_repair(ClientId::new(1)).await;
-        create_and_enqueue_repair_ops(&mut up, 0).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 0);
 
         up.submit_flush(None, None).await;
 
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -1994,7 +1626,7 @@ mod more_tests {
         )
         .await;
 
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -2026,7 +1658,7 @@ mod more_tests {
         )
         .await;
 
-        create_and_enqueue_repair_ops(&mut up, 0).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 0);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -2056,7 +1688,7 @@ mod more_tests {
         up.submit_dummy_read(Block::new_512(1), Buffer::new(512 * 3))
             .await;
 
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
 
@@ -2084,7 +1716,7 @@ mod more_tests {
         up.submit_dummy_read(Block::new_512(2), Buffer::new(512 * 3))
             .await;
 
-        create_and_enqueue_repair_ops(&mut up, 0).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 0);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -2123,8 +1755,8 @@ mod more_tests {
         )
         .await;
 
-        create_and_enqueue_repair_ops(&mut up, 0).await;
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 0);
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -2153,7 +1785,7 @@ mod more_tests {
         up.submit_dummy_read(Block::new_512(0), Buffer::new(512 * 9))
             .await;
 
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -2209,7 +1841,7 @@ mod more_tests {
         // reservation that happens when we need to insert a job like this.
 
         let up = start_up_and_repair(ClientId::new(1)).await;
-        create_and_enqueue_repair_ops(&mut up, 0).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 0);
 
         // A write of blocks 2,3,4 which spans extent 0 and extent 1.
         up.submit_dummy_write(
@@ -2248,7 +1880,7 @@ mod more_tests {
         //   4 |   R R | R     | 3
 
         let up = start_up_and_repair(ClientId::new(1)).await;
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         up.submit_dummy_read(Block::new_512(1), Buffer::new(512 * 3))
             .await;
@@ -2290,7 +1922,7 @@ mod more_tests {
         .await;
 
         // The final repair command
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -2336,7 +1968,7 @@ mod more_tests {
         .await;
 
         // The first repair command
-        create_and_enqueue_repair_ops(&mut up, 0).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 0);
 
         up.submit_dummy_write(
             Block::new_512(4),
@@ -2349,7 +1981,7 @@ mod more_tests {
             .await;
 
         // The second repair command
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
 
         up.submit_dummy_write(
             Block::new_512(5),
@@ -2359,7 +1991,7 @@ mod more_tests {
         .await;
 
         // The third repair command
-        create_and_enqueue_repair_ops(&mut up, 2).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 2);
 
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
@@ -2521,7 +2153,7 @@ mod more_tests {
 
         // Now enqueue the repair on extent 1, it should populate one of the
         // empty job slots.
-        create_and_enqueue_repair_ops(&mut up, 1).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 1);
         let ds = up.downstairs;
         let jobs: Vec<&DownstairsIO> = ds.ds_active().values().collect();
 
@@ -3013,7 +2645,7 @@ mod more_tests {
         drop(ds);
 
         // Put a repair job on the queue.
-        create_and_enqueue_repair_ops(&mut up, 0).await;
+        create_and_enqueue_repair_ops(&mut gw, &mut ds, 0);
 
         // Create a write on extent 1 (not yet repaired)
         up.submit_dummy_write(
